+++ src/components/FirecallItems/FirecallItemDialog.tsx
@@ -1,29 +1,29 @@
+import { ListSubheader } from '@mui/material';
 import Button from '@mui/material/Button';
+import Checkbox from '@mui/material/Checkbox';
 import Dialog from '@mui/material/Dialog';
 import DialogActions from '@mui/material/DialogActions';
 import DialogContent from '@mui/material/DialogContent';
 import DialogContentText from '@mui/material/DialogContentText';
 import DialogTitle from '@mui/material/DialogTitle';
 import FormControl from '@mui/material/FormControl';
+import FormControlLabel from '@mui/material/FormControlLabel';
+import FormGroup from '@mui/material/FormGroup';
 import InputLabel from '@mui/material/InputLabel';
 import MenuItem from '@mui/material/MenuItem';
 import Select, { SelectChangeEvent } from '@mui/material/Select';
 import TextField from '@mui/material/TextField';
+import { StorageReference } from 'firebase/storage';
 import moment from 'moment';
 import React, { useCallback, useState } from 'react';
 import ConfirmDialog from '../dialogs/ConfirmDialog';
 import { FirecallItem } from '../firebase/firestore';
 import MyDateTimePicker from '../inputs/DateTimePicker';
-import { fcItemClasses, fcItemNames, getItemClass } from './elements';
+import FileDisplay from '../inputs/FileDisplay';
+import FileUploader from '../inputs/FileUploader';
+import { fcItemNames, getItemClass } from './elements';
 import { FirecallItemBase } from './elements/FirecallItemBase';
-import { CheckBox } from '@mui/icons-material';
-import FormGroup from '@mui/material/FormGroup';
-import FormControlLabel from '@mui/material/FormControlLabel';
-import Checkbox from '@mui/material/Checkbox';
-import { ListSubheader, Typography } from '@mui/material';
 import { icons } from './elements/icons';
-import FileUploader, { FileDisplay } from '../firebase/storage';
-import { StorageReference } from 'firebase/storage';
 
 export interface FirecallItemDialogOptions {
   onClose: (item?: FirecallItem) => void;
