+++ src/components/inputs/FileUploader.tsx
@@ -3,9 +3,6 @@ import { Typography } from '@mui/material';
 import Button from '@mui/material/Button';
 import { styled } from '@mui/material/styles';
 import {
-  FullMetadata,
-  getDownloadURL,
-  getMetadata,
   getStorage,
   ref,
   StorageReference,
@@ -13,13 +10,11 @@ import {
   UploadMetadata,
   UploadTaskSnapshot,
 } from 'firebase/storage';
-import Image from 'next/image';
-import Link from 'next/link';
-import { useCallback, useEffect, useState } from 'react';
+import { useCallback, useState } from 'react';
 import { v4 as uuid } from 'uuid';
 import { useFirecallId } from '../../hooks/useFirecall';
-import LinearProgressWithLabel from '../inputs/LinearProgressWithLabel';
-import app from './firebase';
+import app from '../firebase/firebase';
+import LinearProgressWithLabel from './LinearProgressWithLabel';
 
 // Initialize Cloud Storage and get a reference to the service
 const storage = getStorage(app);
@@ -150,50 +145,6 @@ export default function FileUploader({
             ))}
         </>
       )}
-    </>
-  );
-}
-
-export interface FileDisplayProps {
-  url: string;
-  showTitleIfImage?: boolean;
-}
-
-export function FileDisplay({
-  url,
-  showTitleIfImage = false,
-}: FileDisplayProps) {
-  const [metadata, setMetadata] = useState<FullMetadata>();
-  const [imageUrl, setImageUrl] = useState<string>();
-
-  useEffect(() => {
-    (async () => {
-      const fileRef = ref(storage, url);
-      const meta = await getMetadata(fileRef);
-      setMetadata(meta);
-      // display image content
-      const downloadUrl = await getDownloadURL(fileRef);
-      setImageUrl(downloadUrl);
-    })();
-  }, [url]);
-  const isImage = metadata?.contentType?.startsWith('image/');
-
-  return (
-    <>
-      <Link href={imageUrl || url} target="_blank">
-        {(!isImage || showTitleIfImage) && (
-          <Typography>{ref(storage, url).name}</Typography>
-        )}
-        {isImage && imageUrl && (
-          <Image
-            src={imageUrl}
-            style={{ maxWidth: 80, maxHeight: 80, margin: 2 }}
-            alt={url}
-            width={80}
-            height={80}
-          />
-        )}
-      </Link>
     </>
   );
 }
